kernel: add automatic fetch & patch dkms tree
Using contrib/dkms/ the source is fetched and patched once (when invoking
patch.sh), which is then copied to /usr/src when invoking "dkms add .".
It is left unchanged, even if other kernels are installed later, in which
case it can fail to build.

Avoid that by calling wget and patch from the module Makefile.  Since we
want all patch files to be available to the dkms build step, add the
Makefile and dkms.conf to the kernel/ subdirectory.

Signed-off-by: Alex Badea <vamposdecampos@gmail.com>